Extract WritableSuite interface
