+++ console-runner/src/main/java/info/javaspec/lang/lambda/WritableSuite.java
@@ -1,0 +1,10 @@
+package info.javaspec.lang.lambda;
+
+import info.javaspec.Spec;
+import info.javaspec.Suite;
+
+public interface WritableSuite extends Suite {
+  void addChildSuite(Suite suite);
+
+  void addSpec(Spec spec);
+}
